Add files via upload
Programming with D3 to calculate the correlation coefficient (r)  which is a measure of the linear relationship between two variables x and y. To get a sense of the connection between the appearance of points -- (x,y) pairs -- in a scatterplot and the value of r, click anywhere on the graph to add points. To remove points, click the Remove Points button and then mouseover points.